ii) Modularized dashboard application in two parts: callbacks and layout. ii) Corrected some route names.
